+++ packages/core/src/EntityManager.ts
@@ -68,6 +68,7 @@ import type {
   Ref,
   RequiredEntityData,
   UnboxArray,
+  NoInfer,
 } from './typings';
 import {
   EventType,
@@ -398,6 +399,7 @@ export class EntityManager<Driver extends IDatabaseDriver = IDatabaseDriver> {
     return where;
   }
 
+  // this method only handles the problem for mongo driver, SQL drivers have their implementation inside QueryBuilder
   protected applyDiscriminatorCondition<Entity extends object>(entityName: string, where: FilterQuery<Entity>): FilterQuery<Entity> {
     const meta = this.metadata.find<Entity>(entityName);
 
@@ -1225,7 +1227,7 @@ export class EntityManager<Driver extends IDatabaseDriver = IDatabaseDriver> {
   async transactional<T>(cb: (em: this) => Promise<T>, options: TransactionOptions = {}): Promise<T> {
     const em = this.getContext(false);
 
-    if (this.disableTransactions) {
+    if (this.disableTransactions || em.disableTransactions) {
       return cb(em);
     }