+++ packages/core/src/utils/Configuration.ts
@@ -186,7 +186,7 @@ export class Configuration<D extends IDatabaseDriver = IDatabaseDriver, EM exten
       this.platform = this.driver.getPlatform();
       this.platform.setConfig(this);
       this.detectSourceFolder(options);
-      this.init();
+      this.init(validate);
     }
   }
 
@@ -337,7 +337,7 @@ export class Configuration<D extends IDatabaseDriver = IDatabaseDriver, EM exten
     this.cache.clear();
   }
 
-  private init(): void {
+  private init(validate: boolean): void {
     if (!this.getMetadataProvider().useCache()) {
       this.options.metadataCache.adapter = NullCacheAdapter;
     }
@@ -354,10 +354,22 @@ export class Configuration<D extends IDatabaseDriver = IDatabaseDriver, EM exten
       this.options.implicitTransactions = this.platform.usesImplicitTransactions();
     }
 
-    const url = this.getClientUrl().match(/:\/\/.*\/([^?]+)/);
+    try {
+      const url = new URL(this.getClientUrl());
 
-    if (url) {
-      this.options.dbName = this.get('dbName', decodeURIComponent(url[1]));
+      if (url.pathname) {
+        this.options.dbName = this.get('dbName', decodeURIComponent(url.pathname).substring(1));
+      }
+    } catch {
+      const url = this.getClientUrl().match(/:\/\/.*\/([^?]+)/);
+
+      if (url) {
+        this.options.dbName = this.get('dbName', decodeURIComponent(url[1]));
+      }
+    }
+
+    if (validate && !this.options.dbName && this.options.clientUrl) {
+      throw new Error("No database specified, `clientUrl` option provided but it's missing the pathname.");
     }
 
     if (!this.options.charset) {