+++ packages/core/src/utils/Utils.ts
@@ -789,8 +789,9 @@ export class Utils {
     stack = stack || new Error().stack!.split('\n');
     // In some situations (e.g. swc 1.3.4+), the presence of a source map can obscure the call to
     // __decorate(), replacing it with the constructor name. To support these cases we look for
-    // Reflect.decorate() as well.
-    let line = stack.findIndex(line => line.match(/__decorate|Reflect\.decorate/));
+    // Reflect.decorate() as well. Also when babel is used, we need to check
+    // the `_applyDecoratedDescriptor` method instead.
+    let line = stack.findIndex(line => line.match(/__decorate|Reflect\.decorate|_applyDecoratedDescriptor/));
 
     // bun does not have those lines at all, only the DecorateProperty/DecorateConstructor,
     // but those are also present in node, so we need to check this only if they weren't found.
@@ -798,7 +799,7 @@ export class Utils {
       // here we handle bun which stack is different from nodejs so we search for reflect-metadata
       const reflectLine = stack.findIndex(line => Utils.normalizePath(line).includes('node_modules/reflect-metadata/Reflect.js'));
 
-      if (reflectLine === -1 || reflectLine + 1 > stack.length || !stack[reflectLine + 1].includes('bun:wrap')) {
+      if (reflectLine === -1 || reflectLine + 2 >= stack.length || !stack[reflectLine + 1].includes('bun:wrap')) {
         return name;
       }
 
@@ -807,10 +808,6 @@ export class Utils {
 
     if (stack[line].includes('Reflect.decorate')) {
       line++;
-    }
-
-    if (stack[line].match(/DecorateProperty|DecorateConstructor/)) {
-      line += 2;
     }
 
     if (Utils.normalizePath(stack[line]).includes('node_modules/tslib/tslib')) {