+++ packages/knex/src/dialects/postgresql/PostgreSqlKnexDialect.ts
@@ -1,5 +1,6 @@
 import type { Configuration } from '@mikro-orm/core';
 import { PostgreSqlTableCompiler } from './PostgreSqlTableCompiler';
+import { PostgreSqlQueryCompiler } from './PostgreSqlQueryCompiler';
 import { MonkeyPatchable } from '../../MonkeyPatchable';
 
 export class PostgreSqlKnexDialect extends MonkeyPatchable.PostgresDialect {
@@ -12,6 +13,11 @@ export class PostgreSqlKnexDialect extends MonkeyPatchable.PostgresDialect {
     tableCompiler.ormConfig = this.ormConfig;
 
     return tableCompiler;
+  }
+
+  queryCompiler() {
+    // eslint-disable-next-line prefer-rest-params
+    return new (PostgreSqlQueryCompiler as any)(this, ...arguments);
   }
 
 }