+++ packages/knex/src/typings.ts
@@ -92,6 +92,7 @@ export interface IndexDef {
   expression?: string; // allows using custom sql expressions
   options?: Dictionary; // for driver specific options
   type?: string | Readonly<{ indexType?: string; storageEngineIndexType?: 'hash' | 'btree'; predicate?: Knex.QueryBuilder }>; // for back compatibility mainly, to allow using knex's `index.type` option (e.g. gin index)
+  deferMode?: DeferMode;
 }
 
 export interface CheckDef<T = unknown> {